// condoplanner-app/src/apiClient/models/UserCondominiumDto.ts
export interface UserCondominiumDto {
    /**
     * 
     * @type {number}
     * @memberof UserCondominiumDto
     */
    userId?: number;
    /**
     * 
     * @type {number}
     * @memberof UserCondominiumDto
     */
    condominiumId?: number;
    /**
     * 
     * @type {ETipoUsuario}
     * @memberof UserCondominiumDto
     */
    userType?: ETipoUsuario;
    /**
     * 
     * @type {boolean}
     * @memberof UserCondominiumDto
     */
    active?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof UserCondominiumDto
     */
    creationTime?: Date;
    /**
     * 
     * @type {Date}
     * @memberof UserCondominiumDto
     */
    endDate?: Date | null;
    /**
     * 
     * @type {UserDto}
     * @memberof UserCondominiumDto
     */
    user?: UserDto;
    /**
     * 
     * @type {CondominiumDto}
     * @memberof UserCondominiumDto
     */
    condominium?: CondominiumDto;
}